requirements.txt
streamlit>=1.30.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-docx>=1.1.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-gigachat>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.7.4
markdown>=3.5
beautifulsoup4>=4.12.0
requests>=2.31.0
httpx>=0.25.0
gigachat>=0.1.0
pandas>=2.1.0
duckduckgo-search>=6.0.0
fpdf2>=2.7.0
